modify:homeBooking component 분리
