+++ src/Pages/HomeBooking/Components/Calendar.js
@@ -10,20 +10,20 @@ import "./Calendar.scss";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faChevronLeft,
-  faChevronRight
+  faChevronRight,
 } from "@fortawesome/free-solid-svg-icons";
 import styled from "styled-components";
 
 const Calendar = ({
+  hideModal,
   setCheckinDate,
   setCheckoutDate,
   dates,
   setDates,
-  hide,
   initialStartDate,
   initialEndDate,
   bedCounts,
-  bathCounts
+  bathCounts,
 }) => {
   const formedStartDate = initialStartDate
     ? moment(2020 + initialStartDate.slice(0, 2) + initialStartDate.slice(4, 6))
@@ -35,7 +35,7 @@ const Calendar = ({
 
   const [dateRange, setdateRange] = useState({
     startDate: formedStartDate,
-    endDate: formedEndDate
+    endDate: formedEndDate,
   });
 
   const [focus, setFocus] = useState("startDate");
@@ -59,13 +59,13 @@ const Calendar = ({
     e.preventDefault();
     setCheckinDate(startDate && startDate.format("YYYY-MM-DD"));
     setCheckoutDate(endDate && endDate.format("YYYY-MM-DD"));
-    hide();
+    hideModal();
   };
 
   return (
     <DateRangePickerWrapper>
       <div className="modalBox">
-        <button className="closeBtn" onClick={hide}>
+        <button className="closeBtn" onClick={hideModal}>
           X
         </button>
         <div className="modalHeader">
